--- src/main/java/flab/SSF/Community/Controller/HomeController.java
@@ -1,13 +1,0 @@
-package flab.SSF.Community.Controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-
-    @GetMapping("/")
-    public String home() {
-        return "home";
-    }
-}
