--- src/main/java/flab/SSF/Community/SsfCommunityApplication.java
@@ -1,13 +1,0 @@
-package flab.SSF.Community;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SsfCommunityApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SsfCommunityApplication.class, args);
-	}
-
-}
