--- src/main/java/flab/SSF/Community/domain/Board.java
@@ -1,92 +1,0 @@
-package flab.SSF.Community.domain;
-
-
-import java.util.Date;
-
-public class Board {
-
-
-    private int no;
-
-    private String title;
-
-    private String content;
-
-    private Date date;
-
-    private int category;
-
-    private int views;
-
-    private String uid;
-
-    public String getUid() {
-        return uid;
-    }
-
-    public void setUid(String uid) {
-        this.uid = uid;
-    }
-
-
-    public int getNo() {
-        return no;
-    }
-
-    public void setNo(int no) {
-        this.no = no;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public Date getDate() {
-        return date;
-    }
-
-    public void setDate(Date date) {
-        this.date = date;
-    }
-
-    public int getCategory() {
-        return category;
-    }
-
-    public void setCategory(int category) {
-        this.category = category;
-    }
-
-    public int getViews() {
-        return views;
-    }
-
-    public void setViews(int views) {
-        this.views = views;
-    }
-
-    @Override
-    public String toString() {
-        return "Board{" +
-                ", no='" + no + '\'' +
-                ", title='" + title + '\'' +
-                ", content='" + content + '\'' +
-                ", date='" + date + '\'' +
-                ", category=" + category +
-                ", views=" + views +
-                ", uid=" + uid +
-                '}';
-    }
-}
