--- src/main/java/flab/SSF/Community/domain/File.java
@@ -1,4 +1,0 @@
-package flab.SSF.Community.domain;
-
-public class File {
-}
