--- src/main/java/flab/SSF/Community/service/CommentService.java
@@ -1,4 +1,0 @@
-package flab.SSF.Community.service;
-
-public class CommentService {
-}
