--- src/main/java/flab/SSF/Community/service/FileService.kt
@@ -1,4 +1,0 @@
-package flab.SSF.Community.service
-
-class FileService {
-}
