--- src/main/java/flab/SSF/Community/service/SearchService.java
@@ -1,4 +1,0 @@
-package flab.SSF.Community.service;
-
-public class SearchService {
-}
