+++ src/main/java/flab/ssf/community/SsfCommunityApplication.java
@@ -1,0 +1,13 @@
+package flab.ssf.community;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SsfCommunityApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SsfCommunityApplication.class, args);
+	}
+
+}
