+++ src/main/java/flab/ssf/community/controller/HomeController.java
@@ -1,0 +1,13 @@
+package flab.ssf.community.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+
+    @GetMapping("/")
+    public String home() {
+        return "home";
+    }
+}
