+++ src/main/java/flab/ssf/community/domain/File.java
@@ -1,0 +1,4 @@
+package flab.ssf.community.domain;
+
+public class File {
+}
