+++ src/main/java/flab/ssf/community/service/CommentService.java
@@ -1,0 +1,4 @@
+package flab.ssf.community.service;
+
+public class CommentService {
+}
