+++ src/main/java/flab/ssf/community/service/FileService.kt
@@ -1,0 +1,4 @@
+package flab.ssf.community.service
+
+class FileService {
+}
