--- src/test/java/flab/SSF/Community/SsfCommunityApplicationTests.java
@@ -1,13 +1,0 @@
-package flab.SSF.Community;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SsfCommunityApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
