+++ src/test/java/flab/ssf/community/SsfCommunityApplicationTests.java
@@ -1,0 +1,13 @@
+package flab.ssf.community;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SsfCommunityApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
